+++ httprunner/models.py
@@ -51,8 +51,8 @@ class TRequest(BaseModel):
     url: Url
     params: Dict[Text, Text] = {}
     headers: Headers = {}
-    req_json: Union[Dict, List] = Field({}, alias="json")
-    data: Union[Text, Dict[Text, Any]] = ""
+    req_json: Union[Dict, List] = Field(None, alias="json")
+    data: Union[Text, Dict[Text, Any]] = None
     cookies: Cookies = {}
     timeout: float = 120
     allow_redirects: bool = True